+++ certbot/certbot/_internal/cli/__init__.py
@@ -101,6 +101,11 @@ def prepare_and_parse_args(plugins, args, detect_defaults=False):
              "flag to 0 disables log rotation entirely, causing "
              "Certbot to always append to the same log file.")
     helpful.add(
+        None, "--preconfigured-renewal", dest="preconfigured_renewal",
+        action="store_true", default=flag_default("preconfigured_renewal"),
+        help=argparse.SUPPRESS
+    )
+    helpful.add(
         [None, "automation", "run", "certonly", "enhance"],
         "-n", "--non-interactive", "--noninteractive",
         dest="noninteractive_mode", action="store_true",